[#I18N] Consolidate locale scripts into a modular i18n CLI
- Replace individual scripts with a central `i18n` package
- Implement subcommands for content, tasks, DB, and validation
- Centralize path configuration and I/O primitives
- Preserve all existing functionality, CLI signatures, and behavior
