--- locales/scripts/__init__.py
@@ -1,1 +1,0 @@
-# Translation service orchestration tooling
