+++ locales/scripts/create-all.sh
