+++ locales/scripts/i18n/commands/__init__.py
@@ -1,0 +1,1 @@
+"""Command groups for the i18n CLI."""
