+++ locales/scripts/i18n/commands/store.py
@@ -1,151 +1,179 @@
-#!/usr/bin/env python3
+"""``db`` command group.
+
+Database management for translation auditing and queries. Ported from the
+legacy ``locales/scripts/store.py``.
+
+The connection + schema lifecycle (``init``/``migrate``) lives in
+:mod:`i18n.db`; this module delegates to those and ports the remaining
+query/export/import logic. Path and table constants come from
+:mod:`i18n.config`.
 """
-Database management for translation auditing and queries.
 
-Architecture:
-- locales/content/{locale}/*.json - Version-controlled source of truth (flat keys)
-- locales/db/tasks.db - Working database for translation workflows
-- locales/db/schema.sql - Table definitions
-
-Tables:
-- translation_tasks: Generated by tasks/create.py for translation workflows
-- glossary: Terminology decisions per locale (committable via SQL export)
-- session_log: Translation session records (committable via SQL export)
-- translation_issues: QC findings requiring review (committable via SQL export)
-
-Usage:
-    python store.py init                          # Create database from schema
-    python store.py migrate                       # Apply schema updates
-    python store.py query "SELECT ..." [--json]   # Run SQL query
-    python store.py export [TABLE]                # Export committable tables to SQL
-    python store.py import [FILE]                 # Import SQL file into database
-"""
+from __future__ import annotations
 
 import argparse
 import hashlib
 import json
 import sqlite3
 import sys
-from contextlib import contextmanager
 from pathlib import Path
-from typing import Any, Iterator, Optional
+from typing import Any, Optional
 
-# Path constants relative to script location
-SCRIPT_DIR = Path(__file__).parent.resolve()
-LOCALES_DIR = SCRIPT_DIR.parent
-DB_DIR = LOCALES_DIR / "db"
-SCHEMA_FILE = DB_DIR / "schema.sql"
-DB_FILE = DB_DIR / "tasks.db"
-
-# Tables that can be exported/imported for version control
-COMMITTABLE_TABLES = ["glossary", "session_log", "translation_issues"]
+from ..config import COMMITTABLE_TABLES, DB_DIR, DB_FILE
+from ..db import get_connection, init_database, migrate_schema
 
 
-@contextmanager
-def get_connection() -> Iterator[sqlite3.Connection]:
-    """Context manager for database connections.
+def register(subparsers) -> None:
+    g = subparsers.add_parser(
+        "db",
+        help="Database management for translation auditing",
+        formatter_class=argparse.RawDescriptionHelpFormatter,
+        epilog="""
+Examples:
+    python store.py init                    # Create database from schema
+    python store.py init --force            # Recreate database
+    python store.py migrate                 # Apply schema updates
+    python store.py query "SELECT * FROM glossary"
+    python store.py query --json "SELECT * FROM session_log"
+    python store.py export                  # Export all committable tables
+    python store.py export glossary         # Export specific table
+    python store.py import                  # Import all SQL files
+    python store.py import glossary.sql     # Import specific file
+        """,
+    )
+    gsub = g.add_subparsers(dest="cmd", required=True)
 
-    Abstracts connection handling for future migration to libsql.
-    """
-    conn = sqlite3.connect(DB_FILE)
-    conn.row_factory = sqlite3.Row
+    # init subcommand
+    init_parser = gsub.add_parser(
+        "init",
+        help="Create database from schema.sql",
+        formatter_class=argparse.RawDescriptionHelpFormatter,
+    )
+    init_parser.add_argument(
+        "--force",
+        "-f",
+        action="store_true",
+        help="Delete existing database and recreate",
+    )
+    init_parser.set_defaults(func=_init)
+
+    # migrate subcommand
+    migrate_parser = gsub.add_parser(
+        "migrate",
+        help="Apply schema updates to existing database",
+        formatter_class=argparse.RawDescriptionHelpFormatter,
+    )
+    migrate_parser.set_defaults(func=_migrate)
+
+    # query subcommand
+    query_parser = gsub.add_parser(
+        "query",
+        help="Run a SQL query",
+        formatter_class=argparse.RawDescriptionHelpFormatter,
+    )
+    query_parser.add_argument("sql", help="SQL query to execute")
+    query_parser.add_argument(
+        "--json", "-j", action="store_true", help="Output results as JSON"
+    )
+    query_parser.set_defaults(func=_query)
+
+    # export subcommand
+    export_parser = gsub.add_parser(
+        "export",
+        help="Export committable tables to SQL files",
+        formatter_class=argparse.RawDescriptionHelpFormatter,
+    )
+    export_parser.add_argument(
+        "table",
+        nargs="?",
+        choices=COMMITTABLE_TABLES,
+        help=f"Table to export (default: all of {', '.join(COMMITTABLE_TABLES)})",
+    )
+    export_parser.set_defaults(func=_export)
+
+    # import subcommand
+    import_parser = gsub.add_parser(
+        "import",
+        help="Import SQL files into database",
+        formatter_class=argparse.RawDescriptionHelpFormatter,
+    )
+    import_parser.add_argument(
+        "file",
+        nargs="?",
+        help="SQL file to import (default: all committable table files)",
+    )
+    import_parser.add_argument(
+        "--no-verify",
+        action="store_true",
+        help="Skip checksum verification",
+    )
+    import_parser.set_defaults(func=_import)
+
+
+# --- handlers -------------------------------------------------------------
+
+
+def _init(args) -> int:
     try:
-        yield conn
-    finally:
-        conn.close()
+        init_database(force=args.force)
+    except FileNotFoundError as e:
+        print(f"Error: {e}", file=sys.stderr)
+        return 1
+    except RuntimeError as e:
+        print(f"Error: {e}", file=sys.stderr)
+        return 1
+    return 0
 
 
-# Schema versions - add new entries when schema changes
-SCHEMA_VERSIONS = [
-    ("001", "initial_tables"),
-    ("002", "translation_tasks"),
-    ("003", "glossary"),
-    ("004", "session_log"),
-    ("005", "source_status"),
-    ("006", "rename_columns"),
-    ("007", "translation_issues"),
-]
+def _migrate(args) -> int:
+    try:
+        migrate_schema()
+    except FileNotFoundError as e:
+        print(f"Error: {e}", file=sys.stderr)
+        return 1
+    except RuntimeError as e:
+        print(f"Error: {e}", file=sys.stderr)
+        return 1
+    return 0
 
 
-def init_database(force: bool = False) -> None:
-    """Initialize database from schema.sql.
-
-    Creates a fresh database with all tables defined in schema.sql.
-
-    Args:
-        force: If True, delete existing database and recreate.
-    """
-    if not SCHEMA_FILE.exists():
-        raise FileNotFoundError(f"Schema file not found: {SCHEMA_FILE}")
-
-    if DB_FILE.exists():
-        if force:
-            DB_FILE.unlink()
-            print(f"Removed existing database: {DB_FILE}")
-        else:
-            print(f"Database already exists: {DB_FILE}")
-            print("Use --force to recreate.")
-            return
-
-    # Ensure db directory exists
-    DB_DIR.mkdir(parents=True, exist_ok=True)
-
-    schema_sql = SCHEMA_FILE.read_text(encoding="utf-8")
-
-    with get_connection() as conn:
-        cursor = conn.cursor()
-        try:
-            cursor.executescript(schema_sql)
-            conn.commit()
-        except sqlite3.Error as e:
-            raise RuntimeError(f"SQL error during schema creation: {e}") from e
-
-    print(f"Created database: {DB_FILE}")
+def _query(args) -> int:
+    try:
+        query(args.sql, output_json=args.json)
+    except FileNotFoundError as e:
+        print(f"Error: {e}", file=sys.stderr)
+        return 1
+    except RuntimeError as e:
+        print(f"Error: {e}", file=sys.stderr)
+        return 1
+    return 0
 
 
-def migrate_schema() -> None:
-    """Apply schema updates to existing database.
+def _export(args) -> int:
+    try:
+        export_tables(table=args.table)
+    except FileNotFoundError as e:
+        print(f"Error: {e}", file=sys.stderr)
+        return 1
+    except RuntimeError as e:
+        print(f"Error: {e}", file=sys.stderr)
+        return 1
+    return 0
 
-    Runs schema.sql which uses CREATE TABLE IF NOT EXISTS,
-    so it safely adds new tables without affecting existing ones.
-    Tracks applied migrations in schema_migrations table.
-    """
-    if not SCHEMA_FILE.exists():
-        raise FileNotFoundError(f"Schema file not found: {SCHEMA_FILE}")
 
-    if not DB_FILE.exists():
-        print(f"Database does not exist: {DB_FILE}")
-        print("Use 'python store.py init' to create it first.")
-        return
+def _import(args) -> int:
+    try:
+        import_tables(file_path=args.file, verify=not args.no_verify)
+    except FileNotFoundError as e:
+        print(f"Error: {e}", file=sys.stderr)
+        return 1
+    except RuntimeError as e:
+        print(f"Error: {e}", file=sys.stderr)
+        return 1
+    return 0
 
-    schema = SCHEMA_FILE.read_text(encoding="utf-8")
 
-    with get_connection() as conn:
-        # Apply schema (idempotent due to IF NOT EXISTS)
-        conn.executescript(schema)
-
-        # Check which versions are already recorded
-        cursor = conn.cursor()
-        cursor.execute("SELECT version FROM schema_migrations")
-        applied = {row[0] for row in cursor.fetchall()}
-
-        # Record any missing versions
-        new_versions = []
-        for version, name in SCHEMA_VERSIONS:
-            if version not in applied:
-                cursor.execute(
-                    "INSERT INTO schema_migrations (version, name) VALUES (?, ?)",
-                    (version, name),
-                )
-                new_versions.append(f"{version}_{name}")
-
-        conn.commit()
-
-    print(f"Schema applied to: {DB_FILE}")
-    if new_versions:
-        print(f"New migrations recorded: {', '.join(new_versions)}")
-    else:
-        print("Schema is up to date.")
+# --- ported logic ---------------------------------------------------------
 
 
 def _load_checksums() -> dict[str, str]:
@@ -427,101 +455,3 @@ def _print_table(rows: list[dict], description: tuple) -> None:
         print(" | ".join(line_parts))
 
     print(f"\n({len(rows)} rows)")
-
-
-def main() -> None:
-    """CLI entry point."""
-    parser = argparse.ArgumentParser(
-        description="Database management for translation auditing.",
-        formatter_class=argparse.RawDescriptionHelpFormatter,
-        epilog="""
-Examples:
-    python store.py init                    # Create database from schema
-    python store.py init --force            # Recreate database
-    python store.py migrate                 # Apply schema updates
-    python store.py query "SELECT * FROM glossary"
-    python store.py query --json "SELECT * FROM session_log"
-    python store.py export                  # Export all committable tables
-    python store.py export glossary         # Export specific table
-    python store.py import                  # Import all SQL files
-    python store.py import glossary.sql     # Import specific file
-        """,
-    )
-
-    subparsers = parser.add_subparsers(dest="command", required=True)
-
-    # init subcommand
-    init_parser = subparsers.add_parser(
-        "init", help="Create database from schema.sql"
-    )
-    init_parser.add_argument(
-        "--force",
-        "-f",
-        action="store_true",
-        help="Delete existing database and recreate",
-    )
-
-    # migrate subcommand
-    subparsers.add_parser(
-        "migrate", help="Apply schema updates to existing database"
-    )
-
-    # query subcommand
-    query_parser = subparsers.add_parser("query", help="Run a SQL query")
-    query_parser.add_argument("sql", help="SQL query to execute")
-    query_parser.add_argument(
-        "--json", "-j", action="store_true", help="Output results as JSON"
-    )
-
-    # export subcommand
-    export_parser = subparsers.add_parser(
-        "export", help="Export committable tables to SQL files"
-    )
-    export_parser.add_argument(
-        "table",
-        nargs="?",
-        choices=COMMITTABLE_TABLES,
-        help=f"Table to export (default: all of {', '.join(COMMITTABLE_TABLES)})",
-    )
-
-    # import subcommand
-    import_parser = subparsers.add_parser(
-        "import", help="Import SQL files into database"
-    )
-    import_parser.add_argument(
-        "file",
-        nargs="?",
-        help="SQL file to import (default: all committable table files)",
-    )
-    import_parser.add_argument(
-        "--no-verify",
-        action="store_true",
-        help="Skip checksum verification",
-    )
-
-    args = parser.parse_args()
-
-    try:
-        if args.command == "init":
-            init_database(force=args.force)
-        elif args.command == "migrate":
-            migrate_schema()
-        elif args.command == "query":
-            query(
-                args.sql,
-                output_json=args.json,
-            )
-        elif args.command == "export":
-            export_tables(table=args.table)
-        elif args.command == "import":
-            import_tables(file_path=args.file, verify=not args.no_verify)
-    except FileNotFoundError as e:
-        print(f"Error: {e}", file=sys.stderr)
-        sys.exit(1)
-    except RuntimeError as e:
-        print(f"Error: {e}", file=sys.stderr)
-        sys.exit(1)
-
-
-if __name__ == "__main__":
-    main()
